Jobs

The section of the package is responsible for handle the jobs.

Index

ModelSelectionGUI.add_job_file

ModelSelectionGUI.clear_current_job

ModelSelectionGUI.clear_jobs_finished

ModelSelectionGUI.clear_jobs_queue

ModelSelectionGUI.consume_job_queue

ModelSelectionGUI.enqueue_job

ModelSelectionGUI.get_current_job

ModelSelectionGUI.get_job

ModelSelectionGUI.get_job

ModelSelectionGUI.get_job_file

ModelSelectionGUI.get_job_finished

ModelSelectionGUI.get_job_queue

ModelSelectionGUI.get_jobs_queue_length

ModelSelectionGUI.job_notify

ModelSelectionGUI.run_job

Contents

ModelSelectionGUI.enqueue_job — Method

enqueue_job(job::ModelSelectionJob)

Enqueues the specified ModelSelectionJob and notifies any waiting tasks.

Arguments

job::ModelSelectionJob: The ModelSelectionJob instance to enqueue.

Returns

The enqueued ModelSelectionJob.

Example

job = ModelSelectionJob(...)  # job is an instance of ModelSelectionJob
enqueue_job(job)

ModelSelectionGUI.consume_job_queue — Method

consume_job_queue()

Continuously consumes jobs from the queue and runs them. This function blocks indefinitely.

Example

consume_job_queue()

ModelSelectionGUI.run_job — Method

run_job(job::ModelSelectionJob)

Runs the specified ModelSelectionJob, reads the necessary data from the job's tempfile, and performs model selection. In case of any exception, sets the job status to FAILED and saves the error message.

Arguments

job::ModelSelectionJob: The ModelSelectionJob instance to run.

Example

job = ModelSelectionJob(...)  # job is an instance of ModelSelectionJob
run_job(job)

ModelSelectionGUI.get_jobs_queue_length — Method

get_jobs_queue_length()

Returns the number of jobs currently enqueued.

Example

get_jobs_queue_length()

ModelSelectionGUI.add_job_file — Method

add_job_file(filehash::String, tempfile::String, filename::String)

Stores a mapping from filehash to the specified tempfile and filename.

Arguments

filehash::String: The file hash.

tempfile::String: The temporary file path.

filename::String: The original file name.

Example

add_job_file("filehash", "tempfile", "filename")

ModelSelectionGUI.get_job_file — Method

get_job_file(filehash::String)

Returns the tempfile and filename associated with the specified filehash.

Arguments

filehash::String: The file hash.

Returns

A Dict with the tempfile and filename.

Example

get_job_file("filehash")

ModelSelectionGUI.job_notify — Function

job_notify(message::String, data::Union{Dict{Any,Any},Nothing} = nothing)

Notifies all subscribers on the default WebSocket channel with the specified message and data.

Arguments

message::String: The message to send.

data::Union{Dict{Any,Any},Nothing}: The data to send.

Example

job_notify("message", Dict("data" => "data"))

ModelSelectionGUI.get_job — Method

get_job(queue::Vector{ModelSelectionJob}, id::String)

Searches for a job with the specified id in the provided queue and returns it. Returns nothing if the job is not found.

Arguments

queue::Vector{ModelSelectionJob}: The queue to search.

id::String: The job id.

Returns

The job with the specified id or nothing if the job is not found.

Example

get_job(queue, "id")

ModelSelectionGUI.get_job — Method

get_job(id::String)

Returns the job with the specified id from the queue of jobs, the current job, or the finished jobs. Returns nothing if the job is not found.

Arguments

id::String: The job id.

Returns

The job with the specified id or nothing if the job is not found.

Example

get_job("id")

ModelSelectionGUI.get_job_queue — Method

get_job_queue(id::String)

Returns the job with the specified id from the queue of jobs. Returns nothing if the job is not found.

Arguments

id::String: The job id.

Returns

The job with the specified id or nothing if the job is not found.

Example

get_job_queue("id")

ModelSelectionGUI.get_current_job — Function

get_current_job(id::Union{String,Nothing} = nothing)

Returns the current job. If id is specified, only returns the current job if its id matches id. Returns nothing if there is no current job or if the id does not match.

Arguments

id::Union{String,Nothing}: The job id.

Returns

The current job or nothing if there is no current job or if the id does not match.

Example

get_current_job()

ModelSelectionGUI.get_job_finished — Method

get_job_finished(id::String)

Returns the job with the specified id from the finished jobs. Returns nothing if the job is not found.

Arguments

id::String: The job id.

Returns

The job with the specified id or nothing if the job is not found.

Example

get_job_finished("id")

ModelSelectionGUI.clear_jobs_queue — Method

clear_jobs_queue()

Clear the jobs queue.

Example

clear_jobs_queue()

ModelSelectionGUI.clear_current_job — Method

clear_current_job()

Clear the current job.

Example

clear_current_job()

ModelSelectionGUI.clear_jobs_finished — Method

clear_jobs_finished()

Clear all the finished jobs.

Example

clear_jobs_finished()