Utils

This section of the documentation is dedicated to explaining the various utility functions used throughout the package.

Index

ModelSelectionGUI.get_csv_filename

ModelSelectionGUI.get_csv_filename

ModelSelectionGUI.get_csv_from_result

ModelSelectionGUI.get_parameters

ModelSelectionGUI.get_pkg_version

ModelSelectionGUI.get_request_filehash

ModelSelectionGUI.get_request_job_id

ModelSelectionGUI.get_txt_filename

ModelSelectionGUI.save_tempfile

Contents

ModelSelectionGUI.get_pkg_version — Method

get_pkg_version(name::AbstractString)

Get the installed version of a Julia package by its name.

Arguments

name::AbstractString: The name of the Julia package.

Returns

The version of the specified package.

Example

get_pkg_version("ModelSelection")

ModelSelectionGUI.save_tempfile — Method

save_tempfile(name::String, data::DataFrame)

Save a DataFrame to a temporary CSV file.

Arguments

name::String: The name of the temporary file.

data::DataFrame: The DataFrame to save.

Returns

The filepath where the DataFrame was saved.

Example

save_tempfile("temp.csv", DataFrame(A = 1:3, B = 4:6))

ModelSelectionGUI.get_csv_filename — Method

get_csv_filename(filename::String, result::ModelSelection.AllSubsetRegression.AllSubsetRegressionResult)

Get a filename for a CSV file that stores the results of an all subset regression.

Arguments

filename::String: The original filename.

result::ModelSelection.AllSubsetRegression.AllSubsetRegressionResult: The result of the all subset regression.

Returns

The filename for the CSV file.

Example

get_csv_filename("data.csv", result)

ModelSelectionGUI.get_csv_filename — Method

get_csv_filename(filename::String, result::ModelSelection.CrossValidation.CrossValidationResult)

Get a filename for a CSV file that stores the results of cross validation.

Arguments

filename::String: The original filename.

result::ModelSelection.CrossValidation.CrossValidationResult: The result of cross validation.

Returns

The filename for the CSV file.

Example

get_csv_filename("data.csv", result)

ModelSelectionGUI.get_txt_filename — Method

get_txt_filename(filename::String)

Get a filename for a TXT file that stores summary data.

Arguments

filename::String: The original filename.

Returns

The filename for the TXT file.

ModelSelectionGUI.get_csv_from_result — Method

get_csv_from_result(filename::String, result::ModelSelectionResult)

Get CSV formatted data from a result of a model selection process.

Arguments

filename::String: The original filename.

result::ModelSelectionResult: The result of the model selection process.

Returns

A dictionary containing the filename and CSV data.

Example

get_csv_from_result("data.csv", result)

ModelSelectionGUI.get_parameters — Method

get_parameters(raw_payload::Dict{String,Any})

Convert a dictionary with string keys to a dictionary with symbol keys.

Arguments

raw_payload::Dict{String,Any}: A dictionary with string keys.

Returns

A dictionary with the same keys and values, but with the keys converted to symbols.

Example

get_parameters(Dict("a" => 1, "b" => 2))

ModelSelectionGUI.get_request_job_id — Method

get_request_job_id(params::Any)

Get the job id from a set of parameters.

Arguments

params::Any: The parameters to search.

Returns

The job id if it exists, nothing otherwise.

Example

get_request_job_id(params)

ModelSelectionGUI.get_request_filehash — Method

get_request_filehash(params::Any)

Get the file hash from a set of parameters.

Arguments

params::Any: The parameters to search.

Returns

The file hash if it exists, nothing otherwise.

Example

get_request_filehash(params)