Variables

This section of the documentation is dedicated to explaining the internal global variables used in this package.

Index

ModelSelectionGUI.ELECTRON_APPLICATION

ModelSelectionGUI.OPEN_CLIENT

ModelSelectionGUI.OPEN_DOCUMENTATION

ModelSelectionGUI.SERVER_HOST

ModelSelectionGUI.SERVER_PORT

ModelSelectionGUI.SSL_ENABLED

Contents

ModelSelectionGUI.SERVER_HOST — Constant

SERVER_HOST

Defines the base URL of the server. By default, it uses the value of SERVER_HOST_DEFAULT.

ModelSelectionGUI.SERVER_PORT — Constant

SERVER_PORT

The port number on which the web server will listen. By default, it uses the value of SERVER_PORT_DEFAULT.

ModelSelectionGUI.SSL_ENABLED — Constant

SSL_ENABLED

Determines whether SSL (Secure Socket Layer) is enabled in the application.

ModelSelectionGUI.OPEN_DOCUMENTATION — Constant

OPEN_DOCUMENTATION

Open a window after starting the server with the documentation. By default, it uses the value of OPEN_DOCUMENTATION_DEFAULT.

ModelSelectionGUI.OPEN_CLIENT — Constant

OPEN_CLIENT

Open a window after starting the server with the client. By default, it uses the value of OPEN_CLIENT_DEFAULT.

ModelSelectionGUI.ELECTRON_APPLICATION — Constant

ELECTRON_APPLICATION

Global variable that holds an instance of an Electron application. It is initialized to nothing, indicating that no instance of the application has been created yet.