Views

This section of the documentation is dedicated to explaining the API views to handle HTTP requests and responses.

Index

ModelSelectionGUI.docs_view

ModelSelectionGUI.estimators_view

ModelSelectionGUI.home_view

ModelSelectionGUI.job_enqueue_view

ModelSelectionGUI.job_info_view

ModelSelectionGUI.job_results_view

ModelSelectionGUI.server_info_view

ModelSelectionGUI.upload_file_view

ModelSelectionGUI.websocket_channel_view

Contents

ModelSelectionGUI.home_view — Method

home_view()

The function handles the home page view. Returns a simple HTML string "ModelSelectionGUI".

Returns

HTTP.Response: A simple HTML string "ModelSelectionGUI".

ModelSelectionGUI.docs_view — Method

docs_view()

The functions render the Swagger UI documentation of the API from the swagger_document data.

Returns

HTTP.Response: A page with the swagger documentation.

ModelSelectionGUI.server_info_view — Method

server_info_view()

Returns information about the server.

Returns

HTTP.Response: A JSON response with the server information.

ModelSelectionGUI.estimators_view — Method

estimators_view()

Get a view of all available estimators. This function returns a view of all available estimators from the ModelSelection.AllSubsetRegression module.

Returns

response::String: A JSON string representation of the estimators.

Returns

HTTP.Response: A JSON response with the server information.

ModelSelectionGUI.upload_file_view — Method

upload_file_view()

Uploads a CSV file to the server. Upon successful upload, the server will process the uploaded file for further operations.

Returns

HTTP.Response: A JSON response with information about the saved file.

ModelSelectionGUI.job_enqueue_view — Method

job_enqueue_view()

Enqueues a model selection job for the specified file. The task will be executed after the previously queued tasks have finished.

Returns

HTTP.Response: A JSON response with information about the created job.

ModelSelectionGUI.job_info_view — Method

job_info_view()

Returns the info of a model selection job. If the job is finished, it also includes the summary of a model selection job.

Returns

HTTP.Response: A JSON response with information about the selected job.

ModelSelectionGUI.job_results_view — Method

jobresultsview()

Returns the result file of a specific type for a model selection job.

Returns

HTTP.Response: A text file with the selected result.

ModelSelectionGUI.websocket_channel_view — Method

websocket_channel_view()

The function handles the view for WebSocket channel subscriptions. It subscribes the client to the default WebSocket channel and returns a confirmation string.

Returns

String: A confirmation string.